==> notebooks/07_examples.ipynb <==
%load_ext autoreload
%autoreload 2

%%time

# make sure we are not oversubscribing CPUs
# by adjusting `n_threads`, `corals.cor_matrix` can be parallelized
from corals.threads import set_threads_for_external_libraries
set_threads_for_external_libraries(n_threads=1)

# Create data

# create data

# imports
import numpy as np

# create random data
n_features = 20000
n_samples = 50
X = np.random.random((n_samples, n_features))

# Full correlation matrix

%%time
# full correlation matrix computation
from corals.correlation.full.default import cor_full
cor_matrix = cor_full(X)

# Top-k correlations

%%time
# top-k correlation matrix computation using Spearman correlation and one job
from corals.correlation.topk.default import cor_topk
cor_topk_result = cor_topk(X, k=0.001, correlation_type="spearman", n_jobs=8)

# Top-k correlation differences

%%time
# top-k differential correlation matrix computation using Spearman correlation and one job
from corals.correlation.topkdiff.default import cor_topkdiff
X1 = X
X2 = np.random.random((n_samples, n_features))
cor_topkdiff_result = cor_topkdiff(X1, X2, k=0.001, correlation_type="spearman", n_jobs=8)
